# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Credit Broker 2.0",
        "description": "Документация\n * Во всех ответах приходит объект: `data`, кроме ошибок: В ошибках `401` имеется только поле `message`; в ошибке `422` имеются поля: `errors и message`.\n * Для авторизованных запросов необходимо вставить в HEADER `Authorization ` Bearer токен полученный при авторизации. Пример:\n *     Authorization: Bearer eyJ0eXAiOiJKV1QiLCJhbM3MjlkIn0.eyJhdWQiOiI2IiwianRpIjoiYThlYjY3NGVjMzIyMmU3MWY5Nzk3MzMwMjU\n * Токен можно получить при входе в систему или регистрации. Сам токен будет расположен в ответе HEADER:\n *      X-Access-Token: eyJ0eXAiOiJKV1QiLCJhbM3MjlkIn0.eyJhdWQiOiI2IiwianRpIjoiYThlYjY3NGVjMzIyMmU3MWY5Nzk3MzMwMjU\n *      X-Access-Token-Expires-At: 2023-11-08 10:13:18\n *  Важно! На всех запросах, кроме DEV - необходимо отправлять в HEADER Platform-Code и Device-Id. Иначе будет приходить  `403 ошибка`\n *      Platform-Code: ios\n *      Device-Id: DFIE3232dD\n *\n * На всех запросах в HEADER ОБЯЗАТЕЛЬНО установить:\n *     Accept: application/json",
        "contact": {
            "email": "develop@tsay.uz"
        },
        "version": "2.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "API LOCAL SERVER"
        },
        {
            "url": "http://{tenant}.localhost",
            "description": "API LOCAL SERVER WITH TENANT",
            "variables": {
                "tenant": {
...
            }
...
    "paths": {
        "/v1/company/products": {
            "get": {
                "tags": [
                    "Товар компании"
                ],
                "summary": "Получение всех товаров компании",
                "operationId": "001550709caee2e19116a7965751af97",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/v1.Company.ProductResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    },
                    "403": {
                        "description": "Контент не доступен"
                    },
                    "404": {
                        "description": "Не найдено"
                    },
                    "422": {
                        "description": "Валидация формы"
                    },
                    "429": {
                        "description": "Бан запросов на 1 минуту"
                    },
                    "500": {
                        "description": "Ошибка сервера"
                    }
                },
                "security": [
                    {
                        "bearerAuth": ""
                    }
                ]
            },
            "post": {
                "tags": [
                    "Товар компании"
                ],
                "summary": "Добавление товара в компанию",
                "operationId": "cfcc97449d453e2e0511f1f02622d264",
                "requestBody": {
                    "description": "Тело запроса на добавление товара в компанию",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "name": "Груша",
                                "barcode": "123456789"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/v1.Company.ProductResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    },
                    "403": {
                        "description": "Контент не доступен"
                    },
                    "404": {
                        "description": "Не найдено"
                    },
                    "422": {
                        "description": "Валидация формы"
                    },
                    "429": {
                        "description": "Бан запросов на 1 минуту"
                    },
                    "500": {
                        "description": "Ошибка сервера"
                    }
                },
                "security": [
                    {
                        "bearerAuth": ""
                    }
                ]
            }
        },
        "/v1/company/products/{id}": {
            "get": {
                "tags": [
                    "Товар компании"
                ],
                "summary": "Просмотр товара в компании",
                "operationId": "bd2e09a90d0df58b71989dd12c799783",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID товара",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/v1.Company.ProductResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    },
                    "403": {
                        "description": "Контент не доступен"
                    },
                    "404": {
                        "description": "Не найдено"
                    },
                    "422": {
                        "description": "Валидация формы"
                    },
                    "429": {
                        "description": "Бан запросов на 1 минуту"
                    },
                    "500": {
                        "description": "Ошибка сервера"
                    }
                },
                "security": [
                    {
                        "bearerAuth": ""
                    }
                ]
            },
            "put": {
                "tags": [
                    "Товар компании"
                ],
                "summary": "Редактирование товара в компании",
                "operationId": "3e78a8a720b30699428aa36244294b66",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID товара",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "Тело запроса на редактирование товара в компании",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "name": "Яблоко",
                                "barcode": "987654321"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/v1.Company.ProductResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    },
                    "403": {
                        "description": "Контент не доступен"
                    },
                    "404": {
                        "description": "Не найдено"
                    },
                    "422": {
                        "description": "Валидация формы"
                    },
                    "429": {
                        "description": "Бан запросов на 1 минуту"
                    },
                    "500": {
                        "description": "Ошибка сервера"
                    }
                },
                "security": [
                    {
                        "bearerAuth": ""
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Товар компании"
                ],
                "summary": "Удаление товара из компании",
                "operationId": "6beaaf57cf4eb00fb34280f2cbad377d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID товара",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/v1.Company.ProductResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    },
                    "403": {
                        "description": "Контент не доступен"
                    },
                    "404": {
                        "description": "Не найдено"
                    },
                    "422": {
                        "description": "Валидация формы"
                    },
                    "429": {
                        "description": "Бан запросов на 1 минуту"
                    },
                    "500": {
                        "description": "Ошибка сервера"
                    }
                },
                "security": [
                    {
                        "bearerAuth": ""
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "v1.Company.ProductResource": {},
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Пример: Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiIsImp0aSI6IjI0OT",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            },
            "Platform-Code": {
                "type": "apiKey",
                "description": "Код платформы",
                "name": "Platform-Code",
                "in": "header"
            },
            "Device-Id": {
                "type": "apiKey",
                "description": "ID устройства",
                "name": "Device-Id",
                "in": "header"
            },
            "Company-slug": {
                "type": "apiKey",
                "description": "Slug компании",
                "name": "Company",
                "in": "header"
            }
        }
    }
}